Integrate `Attributes` with `Person` class

+++ src/main/java/seedu/address/logic/commands/PositionDeleteCommand.java
@@ -20,6 +20,9 @@ public class PositionDeleteCommand extends Command {
 
     public static final String MESSAGE_SUCCESS = "Deleted position: %1$s";
     public static final String MESSAGE_POSITION_NOT_FOUND = "The specified position does not exist in the catalog";
+    public static final String MESSAGE_CANNOT_DELETE_DEFAULT_POSITION =
+            "Cannot delete default position: Unassigned Position";
+    public static final String MESSAGE_POSITION_IN_USE = "Cannot delete position currently assigned to persons";
 
     private final Position toDelete;
 
@@ -35,8 +38,16 @@ public class PositionDeleteCommand extends Command {
     public CommandResult execute(Model model) throws CommandException {
         requireNonNull(model);
 
+        if (toDelete.isDefaultUnassignedPosition()) {
+            throw new CommandException(MESSAGE_CANNOT_DELETE_DEFAULT_POSITION);
+        }
+
         if (!model.hasPosition(toDelete)) {
             throw new CommandException(MESSAGE_POSITION_NOT_FOUND);
+        }
+
+        if (!model.getPersonsMatching(person -> person.getPosition().equals(toDelete)).isEmpty()) {
+            throw new CommandException(MESSAGE_POSITION_IN_USE);
         }
 
         model.deletePosition(toDelete);
@@ -64,4 +75,3 @@ public class PositionDeleteCommand extends Command {
                 .toString();
     }
 }
-
