+++ src/main/java/seedu/address/logic/commands/TeamDeleteCommand.java
@@ -20,6 +20,8 @@ public class TeamDeleteCommand extends Command {
 
     public static final String MESSAGE_SUCCESS = "Deleted team: %1$s";
     public static final String MESSAGE_TEAM_NOT_FOUND = "The specified team does not exist in the catalog";
+    public static final String MESSAGE_CANNOT_DELETE_DEFAULT_TEAM = "Cannot delete default team: Unassigned Team";
+    public static final String MESSAGE_TEAM_IN_USE = "Cannot delete team currently assigned to persons";
 
     private final Team toDelete;
 
@@ -35,8 +37,16 @@ public class TeamDeleteCommand extends Command {
     public CommandResult execute(Model model) throws CommandException {
         requireNonNull(model);
 
+        if (toDelete.isDefaultUnassignedTeam()) {
+            throw new CommandException(MESSAGE_CANNOT_DELETE_DEFAULT_TEAM);
+        }
+
         if (!model.hasTeam(toDelete)) {
             throw new CommandException(MESSAGE_TEAM_NOT_FOUND);
+        }
+
+        if (!model.getPersonsMatching(person -> person.getTeam().equals(toDelete)).isEmpty()) {
+            throw new CommandException(MESSAGE_TEAM_IN_USE);
         }
 
         model.deleteTeam(toDelete);
@@ -64,4 +74,3 @@ public class TeamDeleteCommand extends Command {
                 .toString();
     }
 }
-
